new battle2 logics, QVGA support, update network code, more chages...
git-svn-id: https://fheroes2.svn.sourceforge.net/svnroot/fheroes2/trunk@1400 4f904287-c015-0410-838d-95d3a2656228

+++ fheroes2/src/fheroes2/dialog/dialog_selectcount.cpp
@@ -26,7 +26,7 @@
 #include "button.h"
 #include "dialog.h"
 
-bool Dialog::SelectCount(const std::string &header, u16 min, u16 max, u16 & cur)
+bool Dialog::SelectCount(const std::string &header, u32 min, u32 max, u32 & cur)
 {
     Display & display = Display::Get();
